+++ apps/website/src/components/component-showcase/component-showcase.tsx
@@ -21,6 +21,7 @@ import {
   Card,
   CardBody,
   Heading,
+  Tag,
 } from '@chakra-ui/react'
 import {
   Field,
@@ -81,6 +82,14 @@ export const ComponentShowcase = () => {
     zIndex: 2,
   }
 
+  const menuItemStyles = {
+    bg: 'transparent',
+    _hover: {
+      bg: 'blackAlpha.50',
+      _dark: { bg: 'whiteAlpha.50' },
+    },
+  }
+
   const cardStyles = {
     borderWidth: useColorModeValue(0, '1px'),
     fontSize: 'md',
@@ -114,14 +123,14 @@ export const ComponentShowcase = () => {
         }
         return (
           <OverflowMenu menuListProps={menuStyles} flip={false}>
-            <MenuItem command="D" icon={<FiCopy />}>
+            <MenuItem command="D" icon={<FiCopy />} sx={menuItemStyles}>
               Duplicate
             </MenuItem>
-            <MenuItem command="R" icon={<FiPenTool />}>
+            <MenuItem command="R" icon={<FiPenTool />} sx={menuItemStyles}>
               Rename
             </MenuItem>
             <MenuDivider />
-            <MenuItem command="⌘ ⌫" icon={<FiDelete />}>
+            <MenuItem command="⌘ ⌫" icon={<FiDelete />} sx={menuItemStyles}>
               Delete
             </MenuItem>
           </OverflowMenu>
@@ -133,7 +142,7 @@ export const ComponentShowcase = () => {
   const data: Data[] = [
     {
       id: '1',
-      amount: '30+',
+      amount: '40+',
       name: 'Open Source Components',
     },
   ]
@@ -142,7 +151,7 @@ export const ComponentShowcase = () => {
     <Section
       position="relative"
       innerWidth="container.2xl"
-      height="480px"
+      height="500px"
       overflow="hidden"
     >
       <BackgroundGradient
@@ -150,7 +159,7 @@ export const ComponentShowcase = () => {
         opacity={useColorModeValue(0.2, 0.2)}
       />
       <Box position="relative" my="6">
-        <HStack position="relative">
+        <HStack position="relative" transform="translateX(50px)">
           <Float
             position="absolute"
             top="100px"
@@ -165,20 +174,10 @@ export const ComponentShowcase = () => {
                 data={data}
                 isSelectable
                 isSortable
-                // initialState={{
-                //   selectedRowIds: [0],
-                // }}
-                // sx={{
-                //   th: {
-                //     borderBottomColor: useColorModeValue(
-                //       'blackAlpha.200',
-                //       'whiteAlpha.300'
-                //     ),
-                //   },
-                //   td: {
-                //     borderBottom: 0,
-                //   },
-                // }}
+                initialState={{
+                  /* @ts-ignore */
+                  selectedRowIds: ['1'],
+                }}
               />
             </Card>
           </Float>
@@ -274,7 +273,7 @@ export const ComponentShowcase = () => {
             </Card>
           </Float>
 
-          <Float delay={2} position="absolute" left="860px" top="30px">
+          <Float delay={2} position="absolute" left="920px" top="80px">
             <Card width="320px" {...cardStyles} id="showcase-property-list">
               <Tabs colorScheme="purple">
                 <TabList>
@@ -304,24 +303,10 @@ export const ComponentShowcase = () => {
                             flip={false}
                             autoSelect={false}
                           >
-                            <MenuItemOption
-                              value="Open"
-                              bg="transparent"
-                              _hover={{
-                                bg: 'blackAlpha.50',
-                                _dark: { bg: 'whiteAlpha.50' },
-                              }}
-                            >
+                            <MenuItemOption value="Open" sx={menuItemStyles}>
                               Open
                             </MenuItemOption>
-                            <MenuItemOption
-                              value="Closed"
-                              bg="transparent"
-                              _hover={{
-                                bg: 'blackAlpha.50',
-                                _dark: { bg: 'whiteAlpha.50' },
-                              }}
-                            >
+                            <MenuItemOption value="Closed" sx={menuItemStyles}>
                               Closed
                             </MenuItemOption>
                           </Select>
@@ -341,11 +326,17 @@ export const ComponentShowcase = () => {
                 <Heading size="sm" fontWeight="medium" flex="1">
                   Github
                 </Heading>
-                <Button variant="subtle" colorScheme="green">
+                <Tag variant="subtle" colorScheme="green" rounded="full" px="4">
                   Enabled
-                </Button>
-                <OverflowMenu menuListProps={menuStyles} flip={false}>
-                  <MenuItem>Settings</MenuItem>
+                </Tag>
+                <OverflowMenu
+                  menuListProps={menuStyles}
+                  flip={false}
+                  placement="bottom-end"
+                >
+                  <MenuItem sx={menuItemStyles}>Settings</MenuItem>
+                  <MenuDivider />
+                  <MenuItem sx={menuItemStyles}>Disconnect</MenuItem>
                 </OverflowMenu>
               </HStack>
             </Card>