+++ apps/website/src/docs/components/mobile-nav.tsx
@@ -203,7 +203,7 @@ export const MobileNavButton = React.forwardRef(
     return (
       <IconButton
         ref={ref}
-        display={{ base: 'flex', md: 'none' }}
+        display={{ base: 'flex', lg: 'none' }}
         aria-label="Open menu"
         fontSize="20px"
         color={useColorModeValue('gray.800', 'inherit')}