+++ packages/saas-ui-core/src/error-boundary/error-boundary.tsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import { SaasContext } from '../provider'
 
 export interface ErrorBoundaryProps {
-  errorComponent?: React.ReactNode
+  fallback?: React.ReactNode
   children: React.ReactNode
 }
 
@@ -11,7 +11,11 @@ export interface ErrorBoundaryState {
   error?: Error | null
   errorInfo?: any
 }
-
+/**
+ * A container component that catches errors and displays a fallback UI.
+ *
+ * @see Docs https://saas-ui.dev/docs/components/utils/error-boundary
+ */
 export class ErrorBoundary extends React.Component<ErrorBoundaryProps> {
   state: ErrorBoundaryState
 
@@ -32,7 +36,7 @@ export class ErrorBoundary extends React.Component<ErrorBoundaryProps> {
 
   render() {
     if (this.state.error) {
-      return this.props.errorComponent || <h1>Something went wrong.</h1>
+      return this.props.fallback || <h1>Something went wrong.</h1>
     }
 
     return this.props.children