+++ packages/saas-ui-core/src/loading-overlay/loading-overlay.stories.tsx
@@ -1,10 +1,10 @@
 import { Container, Box, Stack, Text, Button } from '@chakra-ui/react'
 import * as React from 'react'
 
-import { Loader } from './loader'
+import { LoadingOverlay, LoadingSpinner, LoadingText } from './loading-overlay'
 
 export default {
-  title: 'Components/Feedback/Loader',
+  title: 'Components/Feedback/LoadingOverlay',
   decorators: [
     (Story: any) => (
       <Container mt="40px">
@@ -16,22 +16,32 @@ export default {
 
 export const Basic = () => (
   <Stack height="200px">
-    <Loader />
+    <LoadingOverlay>
+      <LoadingSpinner />
+    </LoadingOverlay>
   </Stack>
 )
 
 export const Overlay = () => (
   <Box height="200px" pos="relative" p="4">
     <Text>This content will be overlayed.</Text>
-    <Loader variant="overlay" />
+    <LoadingOverlay variant="overlay">
+      <LoadingSpinner />
+    </LoadingOverlay>
   </Box>
 )
 
-export const Fullscreen = () => <Loader variant="fullscreen" />
+export const Fullscreen = () => (
+  <LoadingOverlay variant="fullscreen">
+    <LoadingSpinner />
+  </LoadingOverlay>
+)
 
 export const WithText = () => (
   <Stack height="200px">
-    <Loader>Loading...</Loader>
+    <LoadingOverlay>
+      <LoadingText>Loading...</LoadingText>
+    </LoadingOverlay>
   </Stack>
 )
 
@@ -49,7 +59,9 @@ export const ExitAnimation = () => {
   return (
     <Box height="200px" pos="relative" p="4">
       <Button onClick={() => setLoading(true)}>Show loader</Button>
-      <Loader variant="fullscreen" isLoading={isLoading} />
+      <LoadingOverlay variant="fullscreen" isLoading={isLoading}>
+        <LoadingSpinner />
+      </LoadingOverlay>
     </Box>
   )
 }