+++ packages/saas-ui-core/src/loading-overlay/loading-overlay.test.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react'
 
 import { render, testStories } from '@saas-ui/test-utils'
-import * as stories from './loader.stories'
+import * as stories from './loading-overlay.stories'
 
 testStories<typeof stories>(stories)