+++ packages/saas-ui-core/src/structured-list/list.tsx
@@ -15,7 +15,7 @@ import {
   Icon,
 } from '@chakra-ui/react'
 
-import { cx, __DEV__ } from '@chakra-ui/utils'
+import { cx } from '@chakra-ui/utils'
 
 const [StylesProvider, useStyles] = createStylesContext('SuiStructuredList')