+++ packages/saas-ui-date-picker/src/types.ts
@@ -1,8 +1,8 @@
-import { DatePickerState } from '@react-stately/datepicker'
 import { RangeValue } from '@react-types/shared'
+import { DateValue, TimeValue } from '@react-types/datepicker'
 
-export type DateValue = DatePickerState['dateValue']
-export type TimeValue = DatePickerState['timeValue']
+export type { DateValue, TimeValue }
+
 export type FormattedValue = DateValue | Date | string | number
 
 export type DateRangeValue = RangeValue<DateValue | null>