+++ packages/saas-ui-forms/src/display-field.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react'
-import { __DEV__ } from '@chakra-ui/utils'
 import { useFormContext } from 'react-hook-form'
 
 import {
@@ -14,7 +13,11 @@ import { FieldProps } from './field'
 export interface DisplayFieldProps
   extends FormControlProps,
     Omit<FieldProps, 'type' | 'label'> {}
-
+/**
+ *
+ *
+ * @see Docs https://saas-ui.dev/
+ */
 export const DisplayField: React.FC<DisplayFieldProps> = ({
   name,
   label,
@@ -31,15 +34,11 @@ export const DisplayField: React.FC<DisplayFieldProps> = ({
   )
 }
 
-if (__DEV__) {
-  DisplayField.displayName = 'DisplayField'
-}
+DisplayField.displayName = 'DisplayField'
 
 export const FormValue: React.FC<{ name: string }> = ({ name }) => {
   const { getValues } = useFormContext()
   return getValues(name) || null
 }
 
-if (__DEV__) {
-  FormValue.displayName = 'FormValue'
-}
+FormValue.displayName = 'FormValue'