+++ packages/saas-ui-forms/src/form.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
 import { chakra, HTMLChakraProps, forwardRef } from '@chakra-ui/react'
-import { cx, runIfFn, __DEV__ } from '@chakra-ui/utils'
+import { cx, runIfFn } from '@chakra-ui/utils'
 
 import {
   useForm,
@@ -24,15 +24,17 @@ import { Field as DefaultField, FieldProps } from './field'
 
 interface FormRenderContext<
   TFieldValues extends FieldValues = FieldValues,
-  TContext extends object = object
+  TContext extends object = object,
+  TFieldTypes extends FieldProps<TFieldValues> = any
 > extends UseFormReturn<TFieldValues, TContext> {
-  Field: React.FC<FieldProps<TFieldValues>>
+  Field: React.FC<TFieldTypes>
 }
 
 interface FormOptions<
   TFieldValues extends FieldValues = FieldValues,
   TContext extends object = object,
-  TSchema = any
+  TSchema = any,
+  TFieldTypes extends FieldProps<TFieldValues> = FieldProps<TFieldValues>
 > {
   /**
    * The form schema, supports Yup, Zod, and AJV.
@@ -57,25 +59,34 @@ interface FormOptions<
   /**
    * The form children, can be a render prop or a ReactNode.
    */
-  children?: MaybeRenderProp<FormRenderContext<TFieldValues, TContext>>
+  children?: MaybeRenderProp<
+    FormRenderContext<TFieldValues, TContext, TFieldTypes>
+  >
 }
 
 export interface FormProps<
   TFieldValues extends FieldValues = FieldValues,
   TContext extends object = object,
-  TSchema = any
+  TSchema = any,
+  TFieldTypes extends FieldProps<TFieldValues> = FieldProps<TFieldValues>
 > extends UseFormProps<TFieldValues, TContext>,
     Omit<
       HTMLChakraProps<'form'>,
       'children' | 'onChange' | 'onSubmit' | 'onError'
     >,
-    FormOptions<TFieldValues, TContext, TSchema> {}
+    FormOptions<TFieldValues, TContext, TSchema, TFieldTypes> {}
 
+/**
+ * The wrapper component provides context, state, and focus management.
+ *
+ * @see Docs https://saas-ui.dev/docs/components/forms/form
+ */
 export const Form = forwardRef(
   <
     TFieldValues extends FieldValues = FieldValues,
     TContext extends object = object,
-    TSchema = any
+    TSchema = any,
+    TFieldTypes extends FieldProps<FieldValues> = FieldProps<FieldValues>
   >(
     props: FormProps<TFieldValues, TContext, TSchema>,
     ref: React.ForwardedRef<HTMLFormElement>
@@ -136,7 +147,7 @@ export const Form = forwardRef(
     }, [methods, onChange])
 
     const Field: React.FC<FieldProps<TFieldValues>> = React.useMemo(
-      () => (props) => <DefaultField<TFieldValues> {...props} />,
+      () => (props) => <DefaultField {...props} />,
       []
     )
 
@@ -159,9 +170,10 @@ export const Form = forwardRef(
 ) as (<
   TFieldValues extends FieldValues,
   TContext extends object = object,
-  TSchema = any
+  TSchema = any,
+  TFieldTypes extends FieldProps<TFieldValues> = FieldProps<TFieldValues>
 >(
-  props: FormProps<TFieldValues, TContext, TSchema> & {
+  props: FormProps<TFieldValues, TContext, TSchema, TFieldTypes> & {
     ref?: React.ForwardedRef<HTMLFormElement>
   }
 ) => React.ReactElement) & {
@@ -186,22 +198,3 @@ export type GetResolver = <
 ) => Promise<ResolverResult<TFieldValues>>
 
 export type GetFieldResolver = (schema: any) => FieldResolver
-
-export interface CreateFormProps {
-  resolver?: GetResolver
-}
-
-export function createForm<Schema = any>({ resolver }: CreateFormProps) {
-  const CreateForm = <
-    TFieldValues extends FieldValues,
-    TContext extends object = object,
-    TSchema extends Schema = Schema
-  >(
-    props: FormProps<TFieldValues, TContext, TSchema>
-  ) => {
-    const { schema, ...rest } = props
-    return <Form resolver={resolver?.(props.schema)} {...rest} />
-  }
-
-  return CreateForm
-}