+++ packages/saas-ui-date-picker/CHANGELOG.md
@@ -30,6 +30,24 @@
 - Updated dependencies [f1e99198]
   - @saas-ui/modals@2.0.0-next.0
 
+## 0.8.7
+
+### Patch Changes
+
+- 83b1c22: Pass DateRangeInput props down to the internal Input.
+
+## 0.8.6
+
+### Patch Changes
+
+- 5026f70: Restore changes from 0.8.1
+
+## 0.8.5
+
+### Patch Changes
+
+- @saas-ui/modals@1.5.5
+
 ## 0.8.4
 
 ### Patch Changes
@@ -55,6 +73,8 @@
 ### Patch Changes
 
 - @saas-ui/modals@1.5.1
+- 5616dec: Fixed issue where value would not update.
+- 5616dec: Add closeOnSelect property.
 
 ## 0.8.0