+++ tooling/test-utils/CHANGELOG.md
@@ -47,6 +47,13 @@
 - Updated dependencies [f3b09191]
   - @saas-ui/react@2.0.0-next.0
 
+## 2.3.5
+
+### Patch Changes
+
+- Updated dependencies [382e095]
+  - @saas-ui/react@1.9.1
+
 ## 2.3.4
 
 ### Patch Changes